pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anything-ai-exploration"
version = "0.2.0"
description = "Anything AI Exploration for Windows 11 with WSL2 (Ubuntu)"
readme = "README.md"
requires-python = ">=3.10,<3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows :: Windows 11",
    "Operating System :: POSIX :: Linux",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Development Status :: 3 - Alpha",
]

dependencies = [
    # Core ML libraries
    "torch~=2.3.0",
    "torchvision~=0.18.0",
    "accelerate~=0.30.0",
    "bitsandbytes~=0.43.0",
    "peft~=0.11.0",
    # Data processing and visualization
    "numpy>=1.23.0",
    "pandas>=1.5.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scikit-learn>=1.2.0",
    "scipy>=1.10.0",
    # HuggingFace ecosystem
    "transformers~=4.41.0",
    "datasets~=2.19.0",
    "tokenizers>=0.13.0",
    "huggingface_hub>=0.16.0",
    # Training utilities
    "tqdm>=4.65.0",
    "tensorboard>=2.12.0",
    "torch-tb-profiler>=0.4.1",
    "wandb>=0.15.0",
    # File formats and configuration
    "toml>=0.10.0",
    "pyarrow>=12.0.0",
    "python-dotenv>=1.0.0",
    # UI and deployment
    "gradio>=3.32.0",
    # Dependencies for system integration
    "packaging>=23.0",
    "Pillow>=9.5.0",
    "pyparsing>=3.0.9",
    "threadpoolctl>=3.1.0",
    "trl~=0.9.4",
]

[project.optional-dependencies]
dev = [
    "black>=23.3.0",
    "pytest>=7.3.1",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
]

[project.urls]
"Homepage" = "https://github.com/your_username/your_repo_name"
"Bug Tracker" = "https://github.com/your_username/your_repo_name/issues" 
